requirements.txt
spotipy
jishaku
discord.py